Update CoreRT with shared array implementation
